cleaned navbar code
